backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic[email]==2.10.4
pydantic-settings==2.10.1
alembic==1.14.0
celery==5.4.0
pytest==8.3.4
pytest-asyncio==0.24.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-multipart==0.0.17
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.28.1
redis==5.2.1
psycopg2-binary==2.9.10